+++ modules/tools.py
@@ -3,6 +3,7 @@ from fabric.widgets.label import Label
 from fabric.widgets.button import Button
 from fabric.utils.helpers import exec_shell_command_async
 import modules.icons as icons
+from gi.repository import Gdk
 from fabric.utils.helpers import get_relative_path, exec_shell_command_async
 SCREENSHOT_SCRIPT = get_relative_path("../scripts/screenshot.sh")
 OCR_SCRIPT = get_relative_path("../scripts/ocr.sh")
@@ -46,25 +47,30 @@ class Toolbox(Box):
             on_clicked=self.ocr,
         )
 
-        self.btn_close = Button(
+
+
+        self.btn_color = Button(
             name="toolbox-button",
-            child=Label(name="button-label", markup=icons.close),
-            # on_clicked=self.close,
+            tooltip_text="Color Picker\nLeft Click: HEX\nMiddle Click: HSV\nRight Click: RGB",
+            child=Label(
+                name="button-bar-label",
+                markup=icons.colorpicker
+            )
         )
-
-
+        self.btn_color.connect("button-press-event", self.colorpicker)
         self.buttons = [
             self.btn_ssregion,
             self.btn_ssfull,
             self.btn_screenrecord,
             self.btn_ocr,
-            self.btn_close,
+            self.btn_color,
         ]
 
         for button in self.buttons:
             self.add(button)
 
         self.show_all()
+
 
     def close_menu(self):
         self.notch.close_notch()
@@ -86,5 +92,13 @@ class Toolbox(Box):
     def ssregion(self, *args):
         exec_shell_command_async(f"bash {SCREENSHOT_SCRIPT} sf")
         self.close_menu()
-    # def close():
-    #     self.close_menu()
+
+    def colorpicker(self, button, event):
+        if event.button == 1:
+            exec_shell_command_async(f"bash {get_relative_path('../scripts/hyprpicker-hex.sh')}")
+        elif event.button == 2:
+            exec_shell_command_async(f"bash {get_relative_path('../scripts/hyprpicker-hsv.sh')}")
+        elif event.button == 3:
+            exec_shell_command_async(f"bash {get_relative_path('../scripts/hyprpicker-rgb.sh')}")
+
+        self.close_menu()
